Save/load text fixed and VWFed

+++ code/iblard/code/font.h
@@ -9,9 +9,8 @@ extern "C" {
 
 	u32 GetLetterWidth(u32 letter);
 
-	u32 GetSentenceWidth(const u16* text, const u16* end);
+	u32 GetSentenceWidth(const u8* text, const u8* end);
 
-	//int GetYForCentering(const char* text, u32 length);
 }
 
 #endif
